refactored auv_send_data and auv_send_ping threads to separate files
